pyproject.toml
[tool.poetry]
name = "projekt"
version = "0.1.0"
description = ""
authors = ["Jan Sernec"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.0.2"
flask = "^2.1.5"
jupyter = "^1.0.0"
ipykernel = "^6.29.3"
matplotlib = "^3.8.4"
flask-cors = "^4.0.0"
pydantic = "^2.6.4"
pytest = "^8.1.1"
openmeteo-requests = "^1.2.0"
requests-cache = "^1.2.0"
retry-requests = "^2.0.0"
pymongo = "^4.7.2"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
